+++ game/233/CS_ScriptCmd.c
@@ -116,7 +116,7 @@ static void CS_ScriptCmd_ReadOpcode_Main(struct CutsceneObj *cs)
 
 	if (metaFlags & CS_OPCODE_META_HAS_ALIGNED_ARG1)
 	{
-		while ((uintptr_t)cursor & 3)
+		while ((u32)cursor & 3)
 		{
 			cursor++;
 		}
